Don't kill the subprocess in tst_qdbusabstractinterface
Rohan was right in e88f9a92b7ab05ea9bc25083de7dee1b67dd673e to stabilize
the test and reset the state, but killing the subprocess is overkill.
All we need is to reset the state in both applications, which includes
disconnecting and reconnecting to the peer, to discard any sent but not
yet received messages.

Change-Id: Ie01392e6e63bd70ef8345217d3fc641ed63c7aba
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>