QColorDialog: Don't loose focus while color picking
On Windows mouse grabbing doesn't work across processes, which
means we're interacting with other windows when picking colors.

Workaround that by having a transparent 1x1 window below the cursor
at all times so we catch the mouse click. Clicking before the window is below
the cursor won't happen because our timer interval is 30ms, so it's quite fast.

It's hacky but it's what we can do for a feature which was very broken on Windows.

Task-number: QTBUG-43663
Change-Id: I295378e033ddc6d9c2230335f0953a727c21e1dd
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>